[PhpStan] convertTo Method not possible for InputQuantityValue
