+++ models/DataObject/Data/AbstractQuantityValue.php
@@ -15,11 +15,9 @@
 
 namespace Pimcore\Model\DataObject\Data;
 
-use InvalidArgumentException;
 use Pimcore;
 use Pimcore\Model\DataObject\OwnerAwareFieldInterface;
 use Pimcore\Model\DataObject\QuantityValue\Unit;
-use Pimcore\Model\DataObject\QuantityValue\UnitConversionService;
 use Pimcore\Model\DataObject\Traits\OwnerAwareFieldTrait;
 
 abstract class AbstractQuantityValue implements OwnerAwareFieldInterface
@@ -78,33 +76,6 @@ abstract class AbstractQuantityValue implements OwnerAwareFieldInterface
         }
 
         return $this->unit;
-    }
-
-    /**
-     * @param Unit|string $unit target unit. if string provided, unit is tried to be found by abbreviation
-     *
-     * @return self
-     *
-     * @throws \Exception
-     */
-    public function convertTo($unit)
-    {
-        if (is_string($unit)) {
-            $unitObject = Unit::getByAbbreviation($unit);
-            if (!$unitObject instanceof Unit) {
-                throw new InvalidArgumentException('Unit with abbreviation "'.$unit.'" does not exist');
-            }
-            $unit = $unitObject;
-        }
-
-        if (!$unit instanceof Unit) {
-            throw new InvalidArgumentException('Please provide unit as '.Unit::class.' object or as string');
-        }
-
-        /** @var UnitConversionService $converter */
-        $converter = Pimcore::getContainer()->get(UnitConversionService::class);
-
-        return $converter->convert($this, $unit);
     }
 
     abstract public function getValue();
